+++ android/src/main/java/com/timeyaa/flutter_editable_charts/LineSetView.java
@@ -61,23 +61,22 @@ public class LineSetView extends LinearLayout {
 
     /**
      * 设置曲线样式
-     *
-     * @param gridBackgroundColor 背景颜色
+     *  @param gridBackgroundColor 背景颜色
      * @param xAxisTextColor      x坐标颜色
      * @param axisLeftTextColor   左轴颜色
      * @param valueTextColor      值颜色
      * @param valueCircleColor    值圆圈颜色
      * @param valueColor          线颜色
      */
-    public void setLineStyle(int gridBackgroundColor, int xAxisTextColor, int axisLeftTextColor,
-                             int valueTextColor, int valueCircleColor, int valueColor) {
-        lineChart.setGridBackgroundColor(gridBackgroundColor);
-        lineChart.getXAxis().setTextColor(xAxisTextColor);
-        lineChart.getAxisLeft().setTextColor(axisLeftTextColor);
-        lineChart.getData().setValueTextColor(valueTextColor);
+    public void setLineStyle(long gridBackgroundColor, long xAxisTextColor, long axisLeftTextColor,
+                             long valueTextColor, long valueCircleColor, long valueColor) {
+        lineChart.setGridBackgroundColor((int) gridBackgroundColor);
+        lineChart.getXAxis().setTextColor((int) xAxisTextColor);
+        lineChart.getAxisLeft().setTextColor((int) axisLeftTextColor);
+        lineChart.getData().setValueTextColor((int) valueTextColor);
 
-        dataSet.setCircleColor(valueCircleColor);
-        dataSet.setColor(valueColor);
+        dataSet.setCircleColor((int) valueCircleColor);
+        dataSet.setColor((int) valueColor);
 
         dataSet.notifyDataSetChanged();
         lineChart.notifyDataSetChanged();
